+++ statsd/buffer.go
@@ -50,7 +50,7 @@ func (b *statsdBuffer) writeGauge(namespace string, globalTags []string, name st
 	return b.validateNewElement(originalBuffer)
 }
 
-func (b *statsdBuffer) writeCount(namespace string, globalTags []string, name string, value int64, tags []string, rate float64, timestamp int64) error {
+func (b *statsdBuffer) writeCount(namespace string, globalTags []string, name string, value float64, tags []string, rate float64, timestamp int64) error {
 	if b.elementCount >= b.maxElements {
 		return errBufferFull
 	}
